chore: move initializeI18 to localization package from utils
