+++ packages/ui/libs/ui/utils/src/index.ts
@@ -2,6 +2,5 @@ export * from './countries';
 export * from './date';
 export * from './formatters';
 export * from './permissions';
-export * from './localization';
 export * from './types';
 export * from './checks';
